Updated ip lookup classes to the new format
